Move pdf downloading logic to AsyncTask
